refactor: domain module 분리
- 도메인 객체 모듈 분리
