+++ module-common/src/main/java/kitchenpos/generic/converter/QuantityConverter.java
@@ -1,7 +1,7 @@
 package kitchenpos.generic.converter;
 
 
-import kitchenpos.generic.Quantity;
+import kitchenpos.domain.Quantity;
 
 import javax.persistence.AttributeConverter;
 import javax.persistence.Converter;
