+++ module-common/src/main/java/kitchenpos/menu/domain/MenuProductRepository.java
@@ -1,8 +1,7 @@
 package kitchenpos.menu.domain;
 
+import kitchenpos.domain.MenuProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
 
 public interface MenuProductRepository extends JpaRepository<MenuProduct, Long> {
 }
