+++ module-common/src/main/java/kitchenpos/menu/domain/MenuRepository.java
@@ -1,5 +1,6 @@
 package kitchenpos.menu.domain;
 
+import kitchenpos.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
