+++ module-common/src/main/java/kitchenpos/menu/dto/MenuProductResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.menu.dto;
 
-import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.domain.MenuProduct;
 import kitchenpos.product.dto.ProductResponse;
 
 import java.util.List;
