+++ module-common/src/main/java/kitchenpos/order/domain/OrderLineItemRepository.java
@@ -1,8 +1,7 @@
 package kitchenpos.order.domain;
 
+import kitchenpos.domain.OrderLineItem;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
 
 public interface OrderLineItemRepository extends JpaRepository<OrderLineItem, Long> {
 }
