+++ module-common/src/main/java/kitchenpos/order/domain/OrderRepository.java
@@ -1,5 +1,6 @@
 package kitchenpos.order.domain;
 
+import kitchenpos.domain.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderRepository extends JpaRepository<Order, Long> {
