+++ module-common/src/main/java/kitchenpos/product/domain/ProductRepository.java
@@ -1,5 +1,6 @@
 package kitchenpos.product.domain;
 
+import kitchenpos.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
