+++ module-common/src/main/java/kitchenpos/product/dto/ProductResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.product.dto;
 
-import kitchenpos.product.domain.Product;
+import kitchenpos.domain.Product;
 
 public class ProductResponse {
     private final Long id;
