+++ module-common/src/main/java/kitchenpos/product/service/ProductService.java
@@ -1,6 +1,6 @@
 package kitchenpos.product.service;
 
-import kitchenpos.product.domain.Product;
+import kitchenpos.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
 import kitchenpos.product.dto.ProductRequest;
 import kitchenpos.product.dto.ProductResponse;
