+++ module-common/src/main/java/kitchenpos/table/domain/OrderTableRepository.java
@@ -1,5 +1,7 @@
 package kitchenpos.table.domain;
 
+import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
