+++ module-common/src/test/java/kitchenpos/table/ui/TableGroupAcceptanceTest.java
@@ -4,13 +4,10 @@ import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import kitchenpos.AcceptanceTest;
-import kitchenpos.order.domain.Order;
-import kitchenpos.order.util.OrderRequestBuilder;
 import kitchenpos.table.dto.TableGroupRequest;
 import kitchenpos.table.dto.TableGroupResponse;
 import kitchenpos.table.util.OrderTableBuilder;
 import kitchenpos.table.util.TableGroupRequestBuilder;
-import kitchenpos.utils.ResponseUtil;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
