+++ module-domain/src/main/java/kitchenpos/domain/MenuProduct.java
@@ -1,9 +1,4 @@
-package kitchenpos.menu.domain;
-
-import kitchenpos.generic.Money;
-import kitchenpos.generic.Quantity;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.product.domain.Product;
+package kitchenpos.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
