+++ module-domain/src/main/java/kitchenpos/domain/OrderTable.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.domain;
+package kitchenpos.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
