+++ module-domain/src/test/java/kitchenpos/domain/MenuTest.java
@@ -1,8 +1,5 @@
-package kitchenpos.menu.domain;
+package kitchenpos.domain;
 
-import kitchenpos.generic.Money;
-import kitchenpos.generic.Quantity;
-import kitchenpos.product.domain.Product;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
